Extract cross-version helpers from version-helpers.ts
Move HEAD and rolling .x tag handling to dedicated cross-version-helpers.ts:
- getMajorVersion (with HEAD support)
- compareVersions (with HEAD and .x support)
- getDockerImage (with HEAD and .x support)

This keeps release tooling clean - these concepts only exist for
cross-version migration testing.
